authenication working, docblocks added, better error handling, working on converting api subroutes to work with built-in promises

+++ packages/coinstac-graphql/src/data/resolvers.js
@@ -1,4 +1,5 @@
 const rethink = require('rethinkdb');
+const Boom = require('boom');
 const GraphQLJSON = require('graphql-type-json');
 const helperFunctions = require('../auth-helpers');
 
@@ -6,9 +7,12 @@ const helperFunctions = require('../auth-helpers');
 const resolvers = {
   JSON: GraphQLJSON,
   Query: {
+    /**
+     * Returns all computations. Checks user permissions retrieved from JWT middleware validateFunc.
+     */
     fetchAllComputations: ({ auth: { credentials: { permissions } } }, _) => {
-      if (!permissions.computations || !permissions.computations.read) {
-        return { error: 'User not permitted'};
+      if (!permissions.computations.read) {
+        return Boom.unauthorized('Action not permitted');
       }
 
       return helperFunctions.getRethinkConnection()
@@ -16,25 +20,29 @@ const resolvers = {
           rethink.table('computations').run(connection)
         )
         .then((cursor) => {
-          return cursor.toArray(function(err, result) {
-            if (err) throw err;
+          return cursor.toArray((err, result) => {
             return result;
           });
         })
-        .catch(console.log)
     },
-    fetchComputationMetadataByName: (_, args) => {
-      return new Promise ((res, rej) =>
-        rethink.table('computations').filter({ meta: { name: args.computationName } })
-          .run(connection, (error, cursor) => {
-            if (error) throw error;
-            return cursor.toArray(function(err, result) {
-              if (err) throw err;
-              console.log(result);
-              res(result[0]);
-            });
-          })
-      )
+    /**
+     * Returns metadata for specific computation name
+     */
+    fetchComputationMetadataByName: ({ auth: { credentials: { permissions } } }, args) => {
+      if (!permissions.computations.read) {
+        return Boom.unauthorized('Action not permitted');
+      }
+
+      return helperFunctions.getRethinkConnection()
+        .then((connection) =>
+          rethink.table('computations').filter({ meta: { name: args.computationName } })
+            .run(connection)
+        )
+        .then((cursor) => {
+          return cursor.toArray((err, result) => {
+            return result[0];
+          });
+        })
     },
     validateComputation: (_, args) => {
       return new Promise();
@@ -53,27 +61,45 @@ const resolvers = {
     },
   },
   Mutation: {
-    addComputation: (_, args) => {
-      return new Promise ((res, rej) =>
-        rethink.table('computations').insert(
-          args.computationSchema,
-          { 
-            conflict: "replace",
-            returnChanges: true,
-          }
+    /**
+     * Add computation to RethinkDB
+     */
+    addComputation: ({ auth: { credentials: { permissions } } }, args) => {
+      if (!permissions.computations.write) {
+        return Boom.unauthorized('Action not permitted');
+      }
+
+      return helperFunctions.getRethinkConnection()
+        .then((connection) =>
+          rethink.table('computations').insert(
+            args.computationSchema,
+            { 
+              conflict: "replace",
+              returnChanges: true,
+            }
+          )
+          .run(connection)
         )
-          .run(connection, (error, result) => {
-            res(result.changes[0].new_val);
-          })
-      )
+        .then((result) => {
+          return result.changes[0].new_val;
+        })
     },
-    removeAllComputations: () => {
-      return new Promise ((res, rej) =>
-        rethink.table('computations').delete()
-          .run(connection, (error, result) => {
-            res(result);
-          })
-      )
+    /**
+     * Remove all computations from DB
+     */
+    removeAllComputations: ({ auth: { credentials: { permissions } } }, _) => {
+      if (!permissions.computations.write) {
+        return Boom.unauthorized('Action not permitted');
+      }
+
+      return helperFunctions.getRethinkConnection()
+        .then((connection) =>
+          rethink.table('computations').delete()
+            .run(connection)
+        )
+        .then((result) => {
+          return result;
+        })
     },
     removeComputation: (_, args) => {
       return new Promise();
